solve redis
